Moved out EnvironmentFactory implementation from AsciidoctorStructureScannerTest.
